Set the low 64 bits of the IV to zero.
This means that a file would have to be 2^68 bytes long before the
counter overflows. Since that is unlikely to happen most clients
can now get away without caring whether the overflow happens at 64 bits
or at 128 bits.

This makes it easier to support android and iOS where the default
implementations of AES-CTR overflow differently.